finished on lesson 78

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -4,14 +4,22 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Http\Resources\AttendeeResource;
+use App\Http\Traits\CanLoadRelationships;
 use App\Models\Attendee;
 use App\Models\Event;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class AttendeeController extends Controller
 {
+    use CanLoadRelationships;
+
+    private array $relations = ['user'];
+
     public function index(Event $event) {
-        $attendees = $event->attendees()->latest();
+        Gate::authorize('viewAny', Attendee::class);
+
+        $attendees = $this->loadRelationships($event->attendees()->latest());
 
         return AttendeeResource::collection(
             $attendees->paginate()
@@ -19,17 +27,21 @@ class AttendeeController extends Controller
     }
 
     public function store(Request $request, Event $event) {
-        $attendee = $event->attendees()->create([
+        Gate::authorize('create', Attendee::class);
+        $attendee = $this->loadRelationships($event->attendees()->create([
            'user_id' => 1
-        ]);
+        ]));
         return new AttendeeResource($attendee);
     }
 
     public function show(Event $event, Attendee $attendee) {
-        return new AttendeeResource($attendee);
+        Gate::authorize('view', Attendee::class);
+        return new AttendeeResource($this->loadRelationships($attendee));
     }
 
-    public function destroy(string $event, Attendee $attendee) {
+    public function destroy(Event $event, Attendee $attendee) {
+        Gate::authorize('delete', Attendee::class);
+
         $attendee->delete();
         return response()->noContent();
     }
