+++ app/Http/Controllers/Api/EventController.php
@@ -7,6 +7,7 @@ use App\Http\Resources\EventResource;
 use App\Http\Traits\CanLoadRelationships;
 use App\Models\Event;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class EventController extends Controller
 {
@@ -19,6 +20,8 @@ class EventController extends Controller
      */
     public function index()
     {
+        Gate::authorize('viewAny', Event::class);
+
         $query = $this->loadRelationships(Event::query());
 
         return EventResource::collection(
@@ -31,6 +34,8 @@ class EventController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create', Event::class);
+
         $event = Event::create([
             ...$request->validate([
                 'name' => 'required|string|max:255',
@@ -38,7 +43,7 @@ class EventController extends Controller
                 'start_time' => 'required|date',
                 'end_time' => 'required|date|after:start_time',
             ]),
-            'user_id' => 1,
+            'user_id' => $request->user()->id,
         ]);
 
         return new EventResource($this->loadRelationships($event));
@@ -49,6 +54,7 @@ class EventController extends Controller
      */
     public function show(Event $event)
     {
+        Gate::authorize('view', Event::class);
         return new EventResource($this->loadRelationships($event));
     }
 
@@ -57,6 +63,8 @@ class EventController extends Controller
      */
     public function update(Request $request, Event $event)
     {
+        Gate::authorize('update', $event);
+
         $event->update(
             $request->validate([
                 'name' => 'sometimes|string|max:255',   //sometimes makes fields optional
@@ -74,8 +82,8 @@ class EventController extends Controller
      */
     public function destroy(Event $event)
     {
+        Gate::authorize('delete', Event::class);
         $event->delete();
-
         return response(status: 204);
     }
 }
